Tratando o erro 404

+++ src/main/java/br/com/learning/aluraforum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.learning.aluraforum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import br.com.learning.aluraforum.controller.dto.DetalhesDoTopicoDto;
 import br.com.learning.aluraforum.controller.form.AtualizacaoTopicoForm;
@@ -56,28 +57,41 @@ public class TopicosController {
 	}
 
 	@GetMapping("/{id}")
-	public DetalhesDoTopicoDto detalhar(@PathVariable Long id){
+	public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id){
 
-	    Topico topico = topicoRepository.getOne(id);
+	    Optional<Topico> topico = topicoRepository.findById(id);
 
-	    return new DetalhesDoTopicoDto(topico);
+	    if(topico.isPresent()){
+			return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+		}
+
+	    return ResponseEntity.notFound().build();
+
     }
 
     @PutMapping("/{id}")
 	@Transactional
     public ResponseEntity<TopicoDto> atualizar(@PathVariable  Long id, @RequestBody @Valid AtualizacaoTopicoForm form){
+		Optional<Topico> optional = topicoRepository.findById(id);
 
-		Topico topico = form.atualizar(id, topicoRepository);
-
-		return ResponseEntity.ok(new TopicoDto(topico));
+		if(optional.isPresent()){
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
+		return ResponseEntity.notFound().build();
 	}
 
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id){
-		topicoRepository.deleteById(id);
 
-		return  ResponseEntity.ok().build();
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()){
+			topicoRepository.deleteById(id);
+			return  ResponseEntity.ok().build();
+		}
+
+		return ResponseEntity.notFound().build();
 	}
 
 }
